requirements.txt
fastapi==0.104.1
ultralytics==8.0.212
torch==2.1.1
numpy==1.26.2
python-multipart==0.0.6
opencv-python==4.8.1.78
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
Pillow==10.1.0
scipy==1.11.3
streamlit==1.28.2
tenacity==8.2.3
urllib3==2.1.0
uvicorn==0.24.0.post1

